refactor : 공통으로 사용할 모듈 분리 (core)
