+++ kitchenpos-core/src/main/java/kitchenpos/common/BaseEntity.java
